extend \char to support full USV range
git-svn-id: http://scripts.sil.org/svn-public/xetex/TRUNK@188 9f0520aa-a7ea-0310-80d7-d2976a5bef54

+++ TeX/texk/web2c/xetexdir/XeTeXLayoutInterface.cpp
@@ -402,7 +402,7 @@ float getGlyphItalCorr(XeTeXLayoutEngine engine, UInt32 glyphID)
 	return rval;
 }
 
-UInt32 mapCharToGlyph(XeTeXLayoutEngine engine, UInt16 charCode)
+UInt32 mapCharToGlyph(XeTeXLayoutEngine engine, UInt32 charCode)
 {
 	return engine->font->mapCharToGlyph(charCode);
 }
